requirements.txt
# Core Agent & Graph
langgraph
langchain
langchain-core
langchain-google-genai

# API & Web
aiohttp
python-dotenv
requests

# YouTube Automation
google-api-python-client
google-auth-oauthlib
google-auth-httplib2

# Media (Optional but recommended for video processing)
moviepy